lower the thresholds on something like 30sigma from the noise measured at the end of 2012 data

+++ DPGAnalysis/EcalTools/python/pulseDump_cfi.py
@@ -21,7 +21,7 @@ pulseDump = cms.EDAnalyzer("PulseDump",
         kTowerRecovered = cms.vstring('kDeadFE'),
         kDead           = cms.vstring('kNoDataNoTP')
         ),
-    minAmplitudeBarrel = cms.double(143),
-    minAmplitudeEndcap = cms.double(83),
+    minAmplitudeBarrel = cms.double(42),
+    minAmplitudeEndcap = cms.double(78),
     pedestalAnalysis = cms.bool(False)
 )
